requirements.txt
# --------- pytorch --------- #
pytorch-lightning==1.9.4
torchinfo
timm
torchmetrics>=0.11.4
einops
# --------- hydra --------- #
hydra-core
hydra-colorlog
hydra-optuna-sweeper
# --------- loggers --------- #
wandb
tensorboardx
# --------- geolibs --------- #
rasterio>=1.2.10,<2.0.0
geopandas>=0.12.2,<1.0.0
fiona>=1.8.22,<2.0.0
shapely
s2sphere==0.2.5
# --------- others --------- #
pyrootutils     # standardizing the project root setup
rich            # beautiful text formatting in terminal
scipy
scikit-learn
pyyaml
tqdm
pandas
jupyterlab
matplotlib
glob2
numpy
captum          # for Integrated Gradients